+++ src/main/java/projectDWT/core/MatrixUtil.java
@@ -4,26 +4,26 @@ import util.Logger;
 public final class MatrixUtil {
     private MatrixUtil(){}
 
-    public static double[][] copy(double[][] m){
+    public static float[][] copy(float[][] m){
         int h = m.length;
         int w = m[0].length;
-        double[][] out = new double[h][w];
+        float[][] out = new float[h][w];
         for (int y = 0; y < h; y++){
             System.arraycopy(m[y], 0, out[y], 0, w);
         }
         return out;
     }
-    public static double maxAbs(double[][] m){
+    public static float maxAbs(float[][] m){
         double max = 0.0;
-        for (double[] row:m){
-            for (double v:row){
-                double a = Math.abs(v);
+        for (float[] row:m){
+            for (float v:row){
+                float a = Math.abs(v);
                 if (a>max) max = a;
             }
         }
-        return max;
+        return (float) max;
     }
-    public static void thresholdInPlace(double[][] m, double thr){
+    public static void thresholdInPlace(float[][] m, float thr){
         int h = m.length;
         int w = m[0].length;
         int zeros = 0;
@@ -32,7 +32,7 @@ public final class MatrixUtil {
         for (int y = 0; y<h; y++){
             for (int x = 0; x<w; x++){
                 if (Math.abs(m[y][x]) < thr){
-                    m[y][x] = 0.0;
+                    m[y][x] = 0.0F;
                     zeros++;
                 }
             }
@@ -40,7 +40,7 @@ public final class MatrixUtil {
         Logger.info("Thresholding done. Zeroed " + zeros + " / " + total + " coefficients");
 
     }
-    public static int clampToByte(double v){
+    public static int clampToByte(float v){
         int x = (int)Math.round(v);
         if (x<0) return 0;
         if (x>255) return 255;
